requirements.txt
confluent-kafka==2.5.0
pandas==2.2.2
sqlalchemy==2.0.31
mysql-connector-python==8.4.0
pyyaml==6.0.1